requirements.txt
# inequalities can be tested interactively at https://www.darius.page/pipdev/
aenum==3.1.11
anndata==0.8.0
biopython>=1.79
black==22.3.0 # should match .pre-commit-config.yaml
cachetools==4.2.2
choosegpu==0.0.1
crowelab-pyir==1.5.0
Cython<0.30,>=0.29
# for dask and distributed:
# see dask requirements from rapids at https://github.com/rapidsai/cudf/blob/branch-22.04/python/dask_cudf/setup.py
# (change branch to match rapids version in our rapidgenomics_cuda.yml)
dask [complete] == 2022.06.1
distributed==2022.06.1
dtreeviz==1.3.3
enum-mixins==0.0.1
enumchoice==0.0.1
extendanything==0.0.1
fastcluster
fastparquet
feature-engine>=1.4.1
fisher >= 0.1.10
# genetools
-e git+https://github.com/maximz/genetools.git@plot-improvements#egg=genetools
jupyter-server-proxy==1.5.0
jupytext==1.11.5 # should match .pre-commit-config.yaml
kdict==0.1.0
llvmlite>=0.38, <0.39
matplotlib==3.5.2
notebooklog==0.0.1
numba==0.55.2
numpy==1.22.0
openpyxl
optuna==2.2.0
pandas >= 1.4.3, < 2.0 # pandas 1.4.3: "Fixed regression in Groupby.transform() and Groupby.agg() failing with engine="numba" when the index was a MultiIndex" (pandas GH issue 46867)
pre-commit
# protobuf should satisfy tensorflow's setup.py "protobuf >= 3.9.2, < 3.20" (https://github.com/tensorflow/tensorflow/issues/56077)
# and cudf (installed by conda) has constraints (actually only in 22.06, not in 22.04) at https://github.com/rapidsai/cudf/blob/branch-22.04/python/cudf/setup.py
protobuf >= 3.9.2, < 3.20
psutil>=5.7.2
pydot>=1.4.1
pydotplus>=2.0.2
pytest
pytest-cov
pytest-lazy-fixture>=0.6.3
pytest-mock
python-slugify==5.0.2
regressout==0.0.1
scanpy [leiden] == 1.9.1
scikit-learn == 1.1.1
scikit-learn-extra==0.2.0
scipy==1.8.1
scratchcache==0.0.1
seaborn>=0.11.2
SentencePiece
sentinels>=1.0.0
sentry-sdk>=1.5.1
statannotations==0.4.3
static-class-property==0.0.1
StratifiedGroupKFoldRequiresGroups==0.0.1
# summarynb
-e git+https://github.com/maximz/summarynb.git@develop#egg=summarynb
tcrdist3>=0.2.2
tensorboard <2.10, >=2.9
tensorflow==2.9.1
transformers==4.4.1
types-cachetools==4.2.9
typing-extensions==4.3.0
umap-learn>=0.5.1
xgboost-label-encoding==0.0.1
xgboost>=1.6.2
